updating login and other small improvements
